fix(booking): update price type to Long, refund flow and inventory service

+++ src/main/java/com/hotel/hotel_backend/dto/response/AdminBookingResponse.java
@@ -12,7 +12,7 @@ public record AdminBookingResponse(
         LocalDate checkIn,
         LocalDate checkOut,
         long nights,
-        Double totalPrice,
+        Long totalPrice,
         BookingStatus status,
         LocalDateTime createdAt
 ) {
