+++ src/main/java/com/hotel/hotel_backend/dto/response/BookingItemResponse.java
@@ -4,5 +4,5 @@ public record BookingItemResponse(
         Long roomTypeId,
         String roomTypeName,
         Integer quantity,
-        Double stayPrice
+        Long stayPrice
 ) {}
